test: Improving integration tests for endpoint method analyzer
